test(fixtures): add user fixtures + create file per handler
