+++ components/Header/Header.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import { SessionProvider } from 'next-auth/react';
 import server from '../../mocks/api';
-import { sessionHandlerNoName, sessionHandlerLoggedOut } from '../../mocks/api/handlers';
+import { session as sessionHandlers } from '../../mocks/api/handlers/';
 import Header from './Header';
 
 describe('Header', () => {
@@ -21,7 +21,7 @@ describe('Header', () => {
   });
 
   test('displays email when user does\'t have a name', async () => {
-    server.use(sessionHandlerNoName);
+    server.use(sessionHandlers.noName);
 
     render(
       <SessionProvider session={undefined}>
@@ -35,7 +35,7 @@ describe('Header', () => {
   });
 
   test('renders when logged out', async () => {
-    server.use(sessionHandlerLoggedOut);
+    server.use(sessionHandlers.loggedOut);
 
     render(
       <SessionProvider session={undefined}>
