+++ mocks/api/handlers/index.ts
@@ -1,0 +1,9 @@
+import * as session from './session';
+
+export {
+  session,
+};
+
+export default [
+  ...session.default,
+];
